ai-agent/requirements.txt
python-dotenv==1.1.1
langchain-core==0.3.72
langchain-google-genai==2.1.8
langchain==0.3.27
pydantic==2.11.7
todoist-api-python
tzdata
